readers: Extract range_tombstone_change_merger
